+++ src/types.ts
@@ -9,6 +9,6 @@ export interface RoborockPlatformConfig extends PlatformConfig {
   skipDevices?: string;
   transientWarningThrottleHours?: number;
   enableMatter?: boolean;
-  enableMatterServiceAreaBeta?: boolean;
+  cloudOnlyMode?: boolean;
   preferCloudForMatterCommands?: boolean;
 }